correction.
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@40750 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/HostDrivers/Support/darwin/sys/dtrace_glue.h
@@ -5,4 +5,4 @@
 typedef struct solaris_cred cred_t;
 typedef struct solaris_kthread kthread_t;
 typedef struct solaris_x86_saved_state x86_saved_state_t;
-typedef uintptr_t model_t;
+typedef unsigned int model_t;
